requirements.txt
numpy
opencv-python
imutils
Flask
Pillow